pyproject.toml
[build-system]
    requires = ["hatchling"]
    build-backend = "hatchling.build"

[project]
    name = "versecbot_personal_bests"
    version = "0.1"
    authors = [{ name = "Museus", email = "versecbot@museus.dev" }]
    description = "VerSecBot plugin to manage personal bests in a channel"
    readme = "README.md"
    requires-python = ">=3.8"
    classifiers = [
        "Programming Language :: Python :: 3",
        "Operating System :: OS Independent",
    ]

    dependencies = [
        "discord-py>=2,<3",
        "pydantic_settings>=2,<3",
        "versecbot_interface@git+https://github.com/Museus/versecbot-interface.git@v0.2",
    ]

    license = "MIT"
    license-files = ["LICEN[CS]E*"]

    [project.urls]
        Homepage = "https://github.com/Museus/versecbot-PersonalBests"
        Issues = "https://github.com/Museus/versecbot-PersonalBests/issues"


    [project.entry-points.'versecbot.plugins']
        a = 'versecbot_personal_bests:PersonalBestsPlugin'

    [project.optional-dependencies]
        qa = ["ruff", "pytest"]

[tool.hatch.metadata]
    allow-direct-references = true
